+++ R/er_opi.R
@@ -25,6 +25,8 @@ er_opi <- function(B_SOILTYPE_AGR,
   # add visual bindings
   code = value_min = value_max = S_ER_TOT = patterns = indicator = choices = NULL
   S_ER = S_AIM = D_OPI = cfOPI = D_OPI_SOIL = D_OPI_WATER = D_OPI_CLIMATE = D_OPI_LANDSCAPE = D_OPI_BIO = D_OPI_TOT = D_OPI_REWARD = D_OPI_SCORE = NULL
+  S_ER_FARM_TOT = D_FS = cfFS = D_OPI_FARM_TOT = NULL
+  
   # Load bbwp_parms
   bbwp_parms <- BBWPC::bbwp_parms
   
@@ -60,85 +62,99 @@ er_opi <- function(B_SOILTYPE_AGR,
                    B_AREA = B_AREA
                    )
   
-  # count the total score per field (scores / ha)
-  dt[, S_ER_TOT := S_ER_SOIL + S_ER_WATER + S_ER_CLIMATE + S_ER_BIODIVERSITY + S_ER_LANDSCAPE]
-  
+  # count the average total score per field, averaged over the eco goals (scores / ha) 
+  dt[, S_ER_FARM_TOT := max((S_ER_SOIL + S_ER_WATER + S_ER_CLIMATE + S_ER_BIODIVERSITY + S_ER_LANDSCAPE),0.001)/5]
+
   # estimate the farm scores
   
     # melt the table to estimate farm score
     dt.farm <- melt(dt,id.vars = c('id','B_AREA','B_SOILTYPE_AGR'),
-                    measure = patterns("S_ER"), 
+                    measure = patterns("S_ER"),
                     value.name = 'S_ER',
                     variable.name = 'indicator')
-    
-    # adjust the current score in the case that there is no minimum needed
-    # be aware, this is done after calculation of the total score
-    dt.farm[B_SOILTYPE_AGR == 'veen' & indicator == 'S_ER_WATER', S_ER := dt.farm.aim$B_CT_WATER]
-    dt.farm[indicator == 'S_ER_LANDSCAPE', S_ER := dt.farm.aim$B_CT_LANDSCAPE]
-    
+
     # estimate the mean score (score / ha) per indicator for the whole farm
     dt.farm <- dt.farm[,list(farmid = 1, S_ER = weighted.mean(x = S_ER, w = B_AREA)),by='indicator']
-    
+
     # the minimum costs required (euro / ha) required for the medals
     if(medalscore=='gold'){mcosts <- 175} else if(medalscore=='silver') {mcosts <- 100} else {mcosts <- 70}
-    
+
     # add the aim
     dt.farm[indicator=='S_ER_SOIL', S_AIM := dt.farm.aim$B_CT_SOIL]
-    dt.farm[indicator=='S_ER_WATER', S_AIM := dt.farm.aim$B_CT_WATER]
+    dt.farm[indicator=='S_ER_WATER', S_AIM := max(dt.farm.aim$B_CT_WATER,0.001)]
     dt.farm[indicator=='S_ER_CLIMATE', S_AIM := dt.farm.aim$B_CT_CLIMATE]
     dt.farm[indicator=='S_ER_BIODIVERSITY', S_AIM := dt.farm.aim$B_CT_BIO]
-    dt.farm[indicator=='S_ER_LANDSCAPE', S_AIM := dt.farm.aim$B_CT_LANDSCAPE]
-    dt.farm[indicator=='S_ER_TOT', S_AIM := sum(dt.farm.aim[,-1])]
+    dt.farm[indicator=='S_ER_LANDSCAPE', S_AIM := max(dt.farm.aim$B_CT_LANDSCAPE,0.001)]
     dt.farm[indicator=='S_ER_REWARD', S_AIM := mcosts]
-    
-    # estimate distance to target
+    dt.farm[indicator=='S_ER_FARM_TOT', S_AIM :=  dt.farm.aim$B_CT_FARM_TOT]
+
+    # estimate distance to target between 0-100
     dt.farm[, D_OPI := round(pmax(0,pmin(100,S_ER * 100 / S_AIM)),0)]
-    
-    # set a weighing factor on the score per indicator
+
+    # set a weighting factor on the score per indicator (only to be used if thresholds for gold, silver and bronze are relative and not absolute?)
     dt.farm[, cfOPI := wf(D_OPI, type="score")]
-    
-    # weighted farm score
+
+    # weighted farm score (based on distance to target, relative)
     dt.farm.score <- dt.farm[,weighted.mean(x = D_OPI, w = cfOPI)]
     dt.farm.score <- round(dt.farm.score)
+
+    # get the farm mean scores for the five indicators, farm total and costs
+    dt.farm.ind.score <- dcast(dt.farm,farmid ~ indicator, value.var = 'S_ER')
+    dt.farm.ind.score[,c('farmid') := NULL]
+    
+    # rename reward column to costs 
+    setnames(dt.farm.ind.score,"S_ER_REWARD","S_ER_COSTS")
+    
+    # round values
+    cols <- colnames(dt.farm.ind.score)[grepl('S_ER',colnames(dt.farm.ind.score))]
+    dt.farm.ind.score[,c(cols) := lapply(.SD,round,0),.SDcols = cols]
+    
     
     # get the distance to target for the five indicators
-    dt.farm.ind.score <- dcast(dt.farm,farmid ~ indicator, value.var = 'D_OPI')
-    dt.farm.ind.score[,c('farmid') := NULL]
+    dt.farm.ind.opi <- dcast(dt.farm,farmid ~ indicator, value.var = 'D_OPI')
+    
     
   # estimate the distance to target for 5 indicators per field
     
-    # estimate the total contribution of a single field to the desired score on farm level
+    # estimate the total contribution of a single field to the desired score on farm level (add 0.001 to water and landscape indicators for in case these are zero)
     dt[, D_OPI_SOIL := S_ER_SOIL * B_AREA / (dt.farm.aim$B_CT_SOIL * sum(B_AREA))]
-    dt[, D_OPI_WATER := S_ER_WATER * B_AREA / (dt.farm.aim$B_CT_WATER * sum(B_AREA))]
+    dt[, D_OPI_WATER := S_ER_WATER * B_AREA / (max(dt.farm.aim$B_CT_WATER,0.001) * sum(B_AREA))]
     dt[, D_OPI_CLIMATE := S_ER_CLIMATE * B_AREA / (dt.farm.aim$B_CT_CLIMATE * sum(B_AREA))]
     dt[, D_OPI_BIO := S_ER_BIODIVERSITY * B_AREA / (dt.farm.aim$B_CT_BIO * sum(B_AREA))]
-    dt[, D_OPI_LANDSCAPE := S_ER_LANDSCAPE * B_AREA / (dt.farm.aim$B_CT_LANDSCAPE * sum(B_AREA))]
-    dt[, D_OPI_TOT := S_ER_TOT * B_AREA / (sum(dt.farm.aim[,-1]) * sum(B_AREA))]
+    dt[, D_OPI_LANDSCAPE := S_ER_LANDSCAPE * B_AREA / (max(dt.farm.aim$B_CT_LANDSCAPE,0.001) * sum(B_AREA))]
+    dt[, D_OPI_FARM_TOT := S_ER_FARM_TOT * B_AREA / (dt.farm.aim$B_CT_FARM_TOT * sum(B_AREA))]
     dt[, D_OPI_REWARD := S_ER_REWARD * B_AREA / (mcosts * sum(B_AREA))]
+    dt[, D_OPI_TOT := S_ER_FARM_TOT * B_AREA / (dt.farm.aim$B_CT_FARM_TOT * sum(B_AREA))]
     
     # melt the table
     dt.field <- melt(dt,id.vars = c('id','B_AREA'),
-               measure = patterns("D_OPI"), 
-               value.name = 'D_OPI',
-               variable.name = 'indicator')
+                        measure = patterns("D_OPI"), 
+                        value.name = 'D_OPI',
+                        variable.name = 'indicator')
    
     # contribution of a single field, optimized between 0 and 100
-    dt.field[,D_OPI_SCORE := round(100 * pmax(0,pmin(1,D_OPI)),0)]
+    dt.field[,D_OPI_SCORE := round(100 * pmax(0,pmin(1,D_OPI)),0)] # s_er_farm_tot
     
     # add a correction for the distance to target for reward (10%)
-    dt.field[, D_OPI_SCORE := round((0.9 * D_OPI_SCORE) + (10 * dt.farm.ind.score$S_ER_REWARD * 0.01))]
+    dt.field[indicator != "D_OPI_FARM_TOT", D_OPI_SCORE := round((0.9 * D_OPI_SCORE) + (10 * dt.farm.ind.opi$S_ER_REWARD * 0.01))] ##s_er_tot
     
     # dcast output
     dt.field <- dcast(dt.field,id~indicator, value.var = 'D_OPI_SCORE')
     
     # rename the column names to scores
     setnames(dt.field, 
-             old = c('id' , 'D_OPI_SOIL', 'D_OPI_WATER', 'D_OPI_CLIMATE', 'D_OPI_BIO', 'D_OPI_LANDSCAPE', 'D_OPI_TOT', 'D_OPI_REWARD'),
-             new = c("field_id","s_er_soil","s_er_water","s_er_climate","s_er_biodiversity","s_er_landscape","s_er_tot","s_er_reward"))
+             old = c('id' , 'D_OPI_SOIL', 'D_OPI_WATER', 'D_OPI_CLIMATE', 'D_OPI_BIO', 
+                     'D_OPI_LANDSCAPE', 'D_OPI_FARM_TOT', 'D_OPI_REWARD', 'D_OPI_TOT'),
+             new = c("field_id","s_er_soil","s_er_water","s_er_climate","s_er_biodiversity",
+                     "s_er_landscape","s_er_farm_tot","s_er_costs","s_er_tot"))
     
   # collect output
-  out <- list(dt.field.ind.score = dt.field,
+  out <- list(
+              # the relative contribution of a single field to the farm objective
+              dt.field.ind.score = dt.field,
+              # the averaged farm score (mean scores / ha, mean costs / ha)
               dt.farm.ind.score = dt.farm.ind.score,
+              # the BBWP score, being overall distance to target
               dt.farm.score = dt.farm.score)
   
   # return value