+++ tests/testthat/test-er.R
@@ -9,7 +9,7 @@ require(testthat)
   # B_AER_CBS = c('LG05','LG14','LG02')
   # B_LU_BBWP = rep('gras_permanent',3)
   # B_LU_BRP = rep(265,3)
-  # B_LU_ARABLE_ER = c(T,T,T)
+  # B_LU_ARABLE_ER = c(F,F,F)
   # B_LU_PRODUCTIVE_ER = c(T,T,T)
   # B_LU_CULTIVATED_ER = c(T,T,T)
   # M_DRAIN = c(TRUE, FALSE, TRUE)
@@ -46,19 +46,19 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
   test_that("check ecoregeling", {
     expect_equal(
       object = names(test),
-      expected = c('farm','fields'))
+      expected = c('farm','fields','farm_thresholds'))
   })
   
   test_that("check ecoregeling", {
     expect_equal(
       object = colnames(test$fields),
       expected = c("field_id","s_er_soil","s_er_water","s_er_climate","s_er_biodiversity",
-                   "s_er_landscape","s_er_tot","s_er_reward","s_er_medal"))
+                   "s_er_landscape","s_er_costs","s_er_farm_tot","s_er_medal","s_er_reward","s_er_tot"))
   })
   
   test_that("check ecoregeling", {
     expect_equal(
-      object = test$fields$s_er_tot,
+      object = test$fields$s_er_farm_tot,
       expected = c(0,0,0),
       tolerance = 0.01)
   })
@@ -66,7 +66,7 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
   test_that("check ecoregeling", {
     expect_equal(
       object = as.character(unlist(test$farm)),
-      expected = c(0,0,0,0,100,0,0,'none'),
+      expected = c(0,0,0,0,0,0,0,'none',0,0),
       tolerance = 0.01)
   })
 
@@ -142,15 +142,15 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
   # run tests on format and output values
   test_that("check ecoregeling", {
     expect_equal(
-      object = test$fields$s_er_tot,
-      expected = c(100,100,100),
+      object = test$fields$s_er_farm_tot,
+      expected = c(88,71,2),
       tolerance = 0.01)
   })
   
   test_that("check ecoregeling", {
     expect_equal(
       object = as.character(unlist(test$farm)),
-      expected = c(100,100,100,100,100,100,100,'gold'),
+      expected = c(31,61,16,77,71,250,56,'gold',175,100),
       tolerance = 0.01)
   })